+++ lib/pub_api_client.dart
@@ -1,3 +1,5 @@
+export 'package:oauth2/oauth2.dart' show Credentials;
+
 export 'src/helpers/google_packages.dart';
 export 'src/models/barrel.dart';
 export 'src/pub_api_client_base.dart';