+++ test/test_utils.dart
@@ -1,0 +1,12 @@
+import 'dart:io';
+
+final kEnvVars = Platform.environment;
+
+/// User Home Path
+String get kUserHome {
+  if (Platform.isWindows) {
+    return kEnvVars['UserProfile']!;
+  } else {
+    return kEnvVars['HOME']!;
+  }
+}